refactor(input/file): move file input code to standalone package
