+++ src/main/java/edu/iastate/ece/sd/sdmay2126/input/file/InputObjects/FBAParametersJSONObject.java
@@ -1,4 +1,4 @@
-package edu.iastate.ece.sd.sdmay2126.input.InputObjects;
+package edu.iastate.ece.sd.sdmay2126.input.file.InputObjects;
 
 import edu.iastate.ece.sd.sdmay2126.application.FBAParameters;
